Add config sub-types + createConfig factory to runtime-config (closes #293)
Adds APIConfig, AuthConfig, StorageConfig, LoggingConfig, ClaudeConfig,
FeaturesConfig, AppsConfig interfaces and createConfig<T>() factory to
@transformotion/runtime-config. Both apps import shared sub-types from the
package; AIConfig stays per-app (BT has wssUrl, pending #295). AppsConfig
unifies to peers: Record<string,string>, replacing budgetTrackerUrl/launchpadUrl.
