+++ apps/stock-analyser/lib/config/index.ts
@@ -6,60 +6,23 @@
  * Client-safe keys use NEXT_PUBLIC_ prefix.
  */
 
-import { selectProvider, normaliseCrossAppUrl } from '@transformotion/runtime-config'
+import {
+  selectProvider,
+  normaliseCrossAppUrl,
+  createConfig,
+  type APIConfig,
+  type AuthConfig,
+  type StorageConfig,
+  type LoggingConfig,
+  type ClaudeConfig,
+  type FeaturesConfig,
+  type AppsConfig,
+} from '@transformotion/runtime-config'
 
-export interface APIConfig {
-  baseURL: string
-  timeout: number
-}
-
-export interface AuthConfig {
-  provider: 'mock' | 'cognito'
-  cognitoUserPoolId?: string
-  cognitoClientId?: string
-  cognitoRegion?: string
-}
-
+// AIConfig stays per-app until #295 resolves the wssUrl drift between SA and BT
 export interface AIConfig {
   provider: 'mock' | 'claude'
   model: string
-}
-
-export interface StorageConfig {
-  provider: 'local' | 'dynamo'
-  dynamoTablePrefix?: string
-  region?: string
-}
-
-export interface LoggingConfig {
-  provider: 'console' | 'cloudwatch'
-  level: 'debug' | 'info' | 'warn' | 'error'
-  cloudwatchLogGroup?: string
-}
-
-export interface ClaudeConfig {
-  /** Base URL for Claude API proxy (AWS API Gateway) */
-  apiUrl: string
-  /** Base URL for polling job status (analysis-cache Lambda) */
-  cacheUrl: string
-  /** Polling interval in ms */
-  pollInterval: number
-  /** Max polling duration in ms */
-  maxPollTime: number
-}
-
-export interface FeaturesConfig {
-  /** Enable debug logging */
-  debugMode: boolean
-}
-
-export interface AppsConfig {
-  /** URL for the Budget Tracker app (cross-app navigation requires full page load) */
-  budgetTrackerUrl: string
-  /** URL for the Launchpad sign-in page; unauthenticated users are redirected here */
-  signInUrl: string
-  /** URL for the signed-out landing page; used by mock-profile signOut handler */
-  signOutUrl: string
 }
 
 export interface AppConfig {
@@ -73,11 +36,17 @@ export interface AppConfig {
   apps: AppsConfig
 }
 
-/**
- * Load configuration from environment variables.
- * Called once at app startup.
- */
-export function loadConfig(): AppConfig {
+export type {
+  APIConfig,
+  AuthConfig,
+  StorageConfig,
+  LoggingConfig,
+  ClaudeConfig,
+  FeaturesConfig,
+  AppsConfig,
+}
+
+function loadConfig(): AppConfig {
   return {
     api: {
       baseURL: process.env.NEXT_PUBLIC_API_URL || '',
@@ -125,24 +94,13 @@ export function loadConfig(): AppConfig {
       debugMode: process.env.NEXT_PUBLIC_DEBUG_MODE === 'true',
     },
     apps: {
-      budgetTrackerUrl: normaliseCrossAppUrl(process.env.NEXT_PUBLIC_BUDGET_URL, '/budget-tracker/'),
       signInUrl: normaliseCrossAppUrl(process.env.NEXT_PUBLIC_SIGNIN_URL, '/sign-in/'),
       signOutUrl: normaliseCrossAppUrl(process.env.NEXT_PUBLIC_SIGNOUT_URL, '/signed-out/'),
+      peers: {
+        'budget-tracker': normaliseCrossAppUrl(process.env.NEXT_PUBLIC_BUDGET_URL, '/budget-tracker/'),
+      },
     },
   }
 }
 
-// Singleton config instance
-let _config: AppConfig | null = null
-
-export function getConfig(): AppConfig {
-  if (!_config) {
-    _config = loadConfig()
-  }
-  return _config
-}
-
-// Reset config (useful for testing)
-export function resetConfig(): void {
-  _config = null
-}
+export const { getConfig, resetConfig } = createConfig(loadConfig)
